refactor(web): route the classroom writers through commitConfigRepoFiles
